+++ adrestus-core/src/main/java/io/Adrestus/core/UnstakingTransaction.java
@@ -2,9 +2,11 @@ package io.Adrestus.core;
 
 import io.Adrestus.crypto.elliptic.ECDSASignatureData;
 
+import java.math.BigDecimal;
+
 public class UnstakingTransaction extends StakingTransaction {
 
-    public UnstakingTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, int blockNumber, String from, String to, double amount, double transactionFee, int nonce, ECDSASignatureData signature, String name, String details, String website, String identity, double commissionRate, String validatorAddress) {
+    public UnstakingTransaction(String hash, TransactionType type, StatusType status, int zoneFrom, int zoneTo, String timestamp, int blockNumber, String from, String to, BigDecimal amount, BigDecimal transactionFee, int nonce, ECDSASignatureData signature, String name, String details, String website, String identity, double commissionRate, String validatorAddress) {
         super(hash, type, status, zoneFrom, zoneTo, timestamp, blockNumber, from, to, amount, transactionFee, nonce, signature, name, details, website, identity, commissionRate, validatorAddress);
     }
